almost done

+++ oop2_ex01/include/Square.h
@@ -6,6 +6,6 @@ class Square : public BasicShape
 public:
 	Square(double x);
 	virtual ~Square() = default;
-	virtual void draw(int factor) override;
+	virtual void draw(double factor) override;
 private:
 };
